Rename state.rs to session_state.rs
