+++ datafusion/core/src/execution/mod.rs
@@ -18,7 +18,7 @@
 //! Shared state for query planning and execution.
 
 pub mod context;
-pub mod state;
+pub mod session_state;
 
 // backwards compatibility
 pub use crate::datasource::file_format::options;
